requirements.txt
gspread==6.0.2
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.122.0
moviepy==1.0.3
